requirements.txt
ultralytics>=8.0.196
opencv-python-headless>=4.8.1.78
numpy>=1.26.4
pandas>=2.1.1
streamlit>=1.36.0
plotly>=5.15.0
pillow>=10.0.0
psutil>=5.9.0
lap>=0.5.12 # Required for YOLO tracking
